EW-1477 Account Service refactoring
