+++ apps/server/src/modules/account/account.module.ts
@@ -4,19 +4,13 @@ import { SystemModule } from '@modules/system';
 import { UserModule } from '@modules/user';
 import { Module } from '@nestjs/common';
 import { ACCOUNT_REPO } from './domain';
-import { AccountServiceDb } from './domain/services/account-db.service';
 import { AccountService } from './domain/services/account.service';
 import { AccountMikroOrmRepo } from './repo';
 import { DeleteUserAccountDataStep } from './saga';
 
 @Module({
 	imports: [SystemModule, LoggerModule, UserModule, SagaModule],
-	providers: [
-		{ provide: ACCOUNT_REPO, useClass: AccountMikroOrmRepo },
-		AccountServiceDb,
-		AccountService,
-		DeleteUserAccountDataStep,
-	],
+	providers: [{ provide: ACCOUNT_REPO, useClass: AccountMikroOrmRepo }, AccountService, DeleteUserAccountDataStep],
 	exports: [AccountService],
 })
 export class AccountModule {}
